+++ src/files.cmake
@@ -12,12 +12,12 @@ tasks/job.cpp
 tasks/thread_pool.cpp
 template.cpp
 ../include/assert.hpp
-../include/containers/guarded_sequence_container.hpp
-../include/containers/iterator.hpp
-../include/containers/view.hpp
 ../include/meta/arguments/argument.hpp
 ../include/meta/arguments/packaged_arguments.hpp
 ../include/meta/cache/lru_cache.hpp
+../include/meta/containers/guarded_sequence_container.hpp
+../include/meta/containers/iterator.hpp
+../include/meta/containers/view.hpp
 ../include/meta/detail/lru_cache.hpp
 ../include/meta/detail/packaged_arguments.hpp
 ../include/meta/detail/safe_queue.hpp
@@ -39,15 +39,13 @@ template.cpp
 ../include/meta/safe_queue.hpp
 ../include/meta/tasks/job.hpp
 ../include/meta/tasks/thread_pool.hpp
+../include/meta/utility/algorithm.hpp
+../include/meta/utility/concepts.hpp
+../include/meta/utility/function_traits.hpp
+../include/meta/utility/lockable.hpp
+../include/meta/utility/reference_counted.hpp
+../include/meta/utility/scope_value.hpp
+../include/meta/utility/type_traits.hpp
 ../include/pimpl.hpp
 ../include/preprocessor.hpp
-../include/utils/concepts.hpp
-../include/utils/function_traits.hpp
-../include/utils/lockable.hpp
-../include/utils/locks.hpp
-../include/utils/reference_counted.hpp
-../include/utils/scope_value.hpp
-../include/utils/type_traits.hpp
-../include/utils/utility.hpp
-../include/utils/vector.hpp
 )
